+++ compilerAst/src/prog8/ast/AstToSourceTextConverter.kt
@@ -320,7 +320,7 @@ class AstToSourceTextConverter(val output: (text: String) -> Unit, val program: 
         when (numLiteral.type) {
             BaseDataType.BOOL -> output(if(numLiteral.number==0.0) "false" else "true")
             BaseDataType.FLOAT -> output(numLiteral.number.toString())
-            else -> output(numLiteral.number.toInt().toString())
+            else -> output(numLiteral.number.toLong().toString())
         }
     }
